+++ src/NzbDrone.Core/Datastore/Migration/016_add_related_works.cs
@@ -1,0 +1,14 @@
+using FluentMigrator;
+using NzbDrone.Core.Datastore.Migration.Framework;
+
+namespace NzbDrone.Core.Datastore.Migration
+{
+    [Migration(16)]
+    public class AddRelatedBooks : NzbDroneMigrationBase
+    {
+        protected override void MainDbUpgrade()
+        {
+            Alter.Table("Books").AddColumn("RelatedBooks").AsString().WithDefaultValue("[]");
+        }
+    }
+}
