+++ src/NzbDrone.Core/MetadataSource/BookInfo/BookInfoResource/BookResource.cs
@@ -1,11 +1,11 @@
 using System;
 using System.Collections.Generic;
 
-namespace NzbDrone.Core.MetadataSource.SkyHook
+namespace NzbDrone.Core.MetadataSource.BookInfo
 {
     public class BookResource
     {
-        public int GoodreadsId { get; set; }
+        public int ForeignId { get; set; }
         public string TitleSlug { get; set; }
         public string Asin { get; set; }
         public string Description { get; set; }
