+++ src/NzbDrone.Core/MetadataSource/BookInfo/BookInfoResource/WorkResource.cs
@@ -1,15 +1,17 @@
 using System;
 using System.Collections.Generic;
 
-namespace NzbDrone.Core.MetadataSource.SkyHook
+namespace NzbDrone.Core.MetadataSource.BookInfo
 {
     public class WorkResource
     {
-        public int GoodreadsId { get; set; }
+        public int ForeignId { get; set; }
         public string Title { get; set; }
         public string TitleSlug { get; set; }
         public string Url { get; set; }
         public DateTime? ReleaseDate { get; set; }
+        public List<string> Genres { get; set; }
+        public List<int> RelatedWorks { get; set; }
         public List<BookResource> Books { get; set; } = new List<BookResource>();
     }
 }
