+++ src/NzbDrone.Core/MetadataSource/IProvideAuthorInfo.cs
@@ -6,7 +6,7 @@ namespace NzbDrone.Core.MetadataSource
 {
     public interface IProvideAuthorInfo
     {
-        Author GetAuthorInfo(string readarrId, bool useCache = true);
+        Author GetAuthorInfo(string readarrId, bool useCache = true, bool includeBooks = true);
         Author GetAuthorAndBooks(string readarrId, double minPopularity = 0);
         HashSet<string> GetChangedAuthors(DateTime startTime);
     }
