+++ src/NzbDrone.Core/MetadataSource/IProvideBookInfo.cs
@@ -7,6 +7,5 @@ namespace NzbDrone.Core.MetadataSource
     public interface IProvideBookInfo
     {
         Tuple<string, Book, List<AuthorMetadata>> GetBookInfo(string id, bool useCache = true);
-        HashSet<string> GetChangedBooks(DateTime startTime);
     }
 }
