--- src/NzbDrone.Core/MetadataSource/SkyHook/SkyHookResource/AuthorResource.cs
@@ -1,7 +1,0 @@
-namespace NzbDrone.Core.MetadataSource.SkyHook
-{
-    public class AuthorResource : BulkResource
-    {
-        public int GoodreadsId { get; set; }
-    }
-}
