--- src/NzbDrone.Core/MetadataSource/SkyHook/SkyHookResource/BookSearchResource.cs
@@ -1,6 +1,0 @@
-namespace NzbDrone.Core.MetadataSource.SkyHook
-{
-    public class BookSearchResource : BulkResource
-    {
-    }
-}
